Extract fakes from ConcatenatingIteratorTest
